refactor(VCheckbox): convert test from cypress to vitest browser
